+++ src/OpenTelemetry.AutoInstrumentation.Native/environment_variables.h
@@ -53,13 +53,6 @@ const WSTRING log_directory = WStr("OTEL_DOTNET_AUTO_LOG_DIRECTORY");
 // https://github.com/dotnet/coreclr/issues/12468
 const WSTRING clr_disable_optimizations = WStr("OTEL_DOTNET_AUTO_CLR_DISABLE_OPTIMIZATIONS");
 
-// Sets whether to intercept method calls when the caller method is inside a
-// domain-neutral assembly. This is dangerous because the integration assembly
-// the user can guarantee that all Application Pools on the system have at most
-// Default is false. Only used in .NET Framework 4.5 and 4.5.1.
-// https://github.com/DataDog/dd-trace-dotnet/pull/671
-const WSTRING domain_neutral_instrumentation = WStr("OTEL_DOTNET_AUTO_DOMAIN_NEUTRAL_INSTRUMENTATION");
-
 // Indicates whether the profiler is running in the context
 // of Azure App Services
 const WSTRING azure_app_services = WStr("OTEL_DOTNET_AUTO_AZURE_APP_SERVICES");