+++ src/OpenTelemetry.AutoInstrumentation/Configuration/TracerSettings.cs
@@ -17,6 +17,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using OpenTelemetry.AutoInstrumentation.Util;
 
 namespace OpenTelemetry.AutoInstrumentation.Configuration;
 // TODO Move settings to more suitable place?
@@ -37,41 +38,11 @@ public class TracerSettings : Settings
         TracesExporter = ParseTracesExporter(source);
         ConsoleExporterEnabled = source.GetBool(ConfigurationKeys.Traces.ConsoleExporterEnabled) ?? false;
 
-        var instrumentations = new Dictionary<string, TracerInstrumentation>();
-        var enabledInstrumentations = source.GetString(ConfigurationKeys.Traces.Instrumentations);
-        if (enabledInstrumentations != null)
-        {
-            foreach (var instrumentation in enabledInstrumentations.Split(Separator))
-            {
-                if (Enum.TryParse(instrumentation, out TracerInstrumentation parsedType))
-                {
-                    instrumentations[instrumentation] = parsedType;
-                }
-                else
-                {
-                    throw new FormatException($"The \"{instrumentation}\" is not recognized as supported trace instrumentation and cannot be enabled");
-                }
-            }
-        }
-        else
-        {
-            instrumentations = Enum.GetValues(typeof(TracerInstrumentation))
-                .Cast<TracerInstrumentation>()
-                .ToDictionary(
-                    key => Enum.GetName(typeof(TracerInstrumentation), key),
-                    val => val);
-        }
-
-        var disabledInstrumentations = source.GetString(ConfigurationKeys.Traces.DisabledInstrumentations);
-        if (disabledInstrumentations != null)
-        {
-            foreach (var instrumentation in disabledInstrumentations.Split(Separator))
-            {
-                instrumentations.Remove(instrumentation);
-            }
-        }
-
-        EnabledInstrumentations = instrumentations.Values.ToList();
+        EnabledInstrumentations = source.ParseEnabledEnumList<TracerInstrumentation>(
+            enabledConfiguration: ConfigurationKeys.Traces.Instrumentations,
+            disabledConfiguration: ConfigurationKeys.Traces.DisabledInstrumentations,
+            separator: Separator,
+            error: "The \"{0}\" is not recognized as supported trace instrumentation and cannot be enabled");
 
         var providerPlugins = source.GetString(ConfigurationKeys.Traces.ProviderPlugins);
         if (providerPlugins != null)
@@ -102,7 +73,6 @@ public class TracerSettings : Settings
 
         TraceEnabled = source.GetBool(ConfigurationKeys.Traces.Enabled) ?? true;
         LoadTracerAtStartup = source.GetBool(ConfigurationKeys.Traces.LoadTracerAtStartup) ?? true;
-        Integrations = new IntegrationSettingsCollection(source);
     }
 
     /// <summary>
@@ -146,11 +116,6 @@ public class TracerSettings : Settings
     /// Gets the list of legacy sources to be added to the tracer at the startup.
     /// </summary>
     public IList<string> LegacySources { get; } = new List<string>();
-
-    /// <summary>
-    /// Gets a collection of <see cref="Integrations"/> keyed by integration name.
-    /// </summary>
-    public IntegrationSettingsCollection Integrations { get; }
 
     internal static TracerSettings FromDefaultSources()
     {