+++ test/IntegrationTests/AspNetTests.cs
@@ -74,6 +74,9 @@ public class AspNetTests : TestHelper
     [Trait("Containers", "Windows")]
     public async Task SubmitMetrics()
     {
+        // Helps to reduce noise by enabling only AspNet metrics.
+        SetEnvironmentVariable("OTEL_DOTNET_AUTO_METRICS_ENABLED_INSTRUMENTATIONS", "AspNet");
+
         Assert.True(EnvironmentTools.IsWindowsAdministrator(), "This test requires Windows Administrator privileges.");
 
         const int expectedMetricRequests = 1;